pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentledger"
version = "0.1.0"
description = "Workflow identity, attribution, and cost tracking for multi-agent AI systems"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "langgraph>=0.2.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-api>=1.24.0",
    "pyjwt>=2.8.0",
    "pydantic>=2.0.0",
    "rich>=13.7.0",          # pretty terminal output for demo
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
clickhouse = [
    "clickhouse-connect>=0.7.0",  # Phase 2: ClickHouse cost sink
]
policy = [
    "requests>=2.31.0",           # Phase 3: OPA HTTP calls
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agentledger*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
